refactor: move getFunctions to store-sync
