+++ packages/store-sync/package.json
@@ -18,6 +18,7 @@
     "./recs": "./dist/recs/index.js",
     "./sqlite": "./dist/sqlite/index.js",
     "./trpc-indexer": "./dist/trpc-indexer/index.js",
+    "./world": "./dist/world/index.js",
     "./zustand": "./dist/zustand/index.js"
   },
   "typesVersions": {
@@ -45,6 +46,9 @@
       ],
       "trpc-indexer": [
         "./dist/trpc-indexer/index.d.ts"
+      ],
+      "world": [
+        "./dist/world/index.d.ts"
       ],
       "zustand": [
         "./dist/zustand/index.d.ts"
@@ -96,7 +100,6 @@
     "@types/node": "20.12.12",
     "@types/sql.js": "^1.4.4",
     "@viem/anvil": "^0.0.7",
-    "mock-game-contracts": "workspace:*",
     "tsup": "^6.7.0",
     "vitest": "0.34.6"
   },
