+++ packages/store-sync/src/getRecords.ts
@@ -39,7 +39,10 @@ export async function getRecords<table extends Table>(
         indexerUrl: options.indexerUrl,
         filters: [{ tableId: options.table.tableId }],
       });
-      return logs?.logs ?? [];
+      // By default, the indexer includes the `store.Tables` table as part of the snapshot.
+      // Once we change this default, we can remove the filter here.
+      // See https://github.com/latticexyz/mud/issues/3386.
+      return logs?.logs.filter((log) => log.args.tableId === options.table.tableId) ?? [];
     } else {
       debug("fetching records for", options.table.label, "via RPC from", options.rpcUrl);
       const client = createPublicClient({
