+++ packages/store-sync/src/world/common.ts
@@ -1,0 +1,9 @@
+import { Hex } from "viem";
+
+export type WorldFunction = {
+  readonly signature: string;
+  readonly selector: Hex;
+  readonly systemId: Hex;
+  readonly systemFunctionSignature: string;
+  readonly systemFunctionSelector: Hex;
+};
