+++ packages/store-sync/src/world/functionSignatureToAbiItem.ts
