+++ packages/store-sync/src/world/getWorldAbi.test.ts
