+++ packages/store-sync/src/world/index.ts
@@ -1,0 +1,2 @@
+export * from "./getFunctions";
+export * from "./getWorldAbi";
