+++ packages/store-sync/test/mockGame.ts
@@ -1,8 +1,8 @@
 import { execa } from "execa";
 import { anvilRpcUrl } from "./common";
 import { Hex, isHex } from "viem";
-import config from "mock-game-contracts/mud.config";
-import worldAbi from "mock-game-contracts/out/IWorld.sol/IWorld.abi.json";
+import config from "../../../test/mock-game-contracts/mud.config";
+import worldAbi from "../../../test/mock-game-contracts/out/IWorld.sol/IWorld.abi.json";
 
 export { config, worldAbi };
